Move useUpdateEffect to hooks directory
