+++ client/src/hooks/useUpdateEffect.ts
@@ -1,0 +1,16 @@
+import { useRef, useEffect } from 'react';
+
+// https://stackoverflow.com/a/55075818/1526448
+const useUpdateEffect = (effect: Function, dependencies: any[] = []) => {
+  const isInitialMount = useRef(true);
+
+  useEffect(() => {
+    if (isInitialMount.current) {
+      isInitialMount.current = false;
+    } else {
+      effect();
+    }
+  }, dependencies);
+};
+
+export default useUpdateEffect;
